third_party/blink/web_tests/wpt_internal/css/css-contain/container-queries/container-type-sticky-computed.html

<!DOCTYPE html>
<meta charset="utf-8">
<title>CSS Containment Test: Computed values of container-type with sticky</title>
<link rel="help" href="https://drafts.csswg.org/css-contain-3/#container-type">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/css/support/computed-testcommon.js"></script>
<script src="/css/css-contain/container-queries/support/cq-testcommon.js"></script>
<div id="target"></div>
<script>
  setup(() => assert_implements_container_queries());

  test_computed_value('container-type', 'sticky');
  test_computed_value('container-type', 'sticky size', 'size sticky');
  test_computed_value('container-type', 'inline-size sticky');
</script>
